Multiplayer: Connection Scene Flow
Upon clicking on the Play button in the Main Menu scene, the players come into a lobby prototype scene where they can either create a game (join as a host) or join an existing game (join as a client). If a game has already started or there are already 4 players connected or the connection times out, the joining is not approved and the player is showed an appropriate message explaining why they couldn't join.

Players who joined are moved to the character selection scene (prototype). After all players mark themselves as ready by clicking on the Ready button, the game starts.